requirements.txt
torch
torchvision
pytorch-lightning>=1.5.1
editdistance
torchmetrics
pandas
opencv-python
albumentations
numpy
sklearn
scikit-learn
Pillow
matplotlib
learn2learn
